+++ Assets/Scripts/Tools/CameraFeedback.cs
@@ -31,21 +31,18 @@ namespace VRtist
         public Transform vrCamera;
         public Transform rig;
         private GameObject cameraPlane;
-        private GameObject feedbackCamera = null;
 
-        private void Awake()
+        private void Start()
         {
-            CameraManager.Instance.onActiveCameraChanged.AddListener(OnCameraChanged);
             Assert.IsTrue(transform.GetChild(0).name == "CameraFeedbackPlane");
             cameraPlane = transform.GetChild(0).gameObject;
             cameraPlane.GetComponent<MeshRenderer>().material.SetTexture("_UnlitColorMap", CameraManager.EmptyTexture);
+            CameraManager.Instance.RegisterScreen(cameraPlane.GetComponent<MeshRenderer>().material);
+            UpdateTransform();
         }
 
-        protected void Update()
+        public void UpdateTransform()
         {
-            if (!gameObject.activeSelf)
-                return;
-
             Camera cam = CameraManager.Instance.GetActiveCameraComponent();
             float aspect = cam == null ? 16f / 9f : cam.aspect;
             float far = Camera.main.farClipPlane * GlobalState.WorldScale * 0.7f;
@@ -55,27 +52,6 @@ namespace VRtist
             transform.localPosition = direction.normalized * far;
             transform.localRotation = Quaternion.LookRotation(-direction) * Quaternion.Euler(0, 180, 0);
             transform.localScale = new Vector3(scale * aspect, scale, scale);
-        }
-
-        void OnCameraChanged(GameObject _, GameObject activeCamera)
-        {
-            SetActiveCamera(activeCamera);
-        }
-
-        private void SetActiveCamera(GameObject activeCamera)
-        {
-            if (feedbackCamera == activeCamera)
-                return;
-            feedbackCamera = activeCamera;
-            Camera cam = CameraManager.Instance.GetActiveCameraComponent();
-            if (null != cam)
-            {
-                cameraPlane.GetComponent<MeshRenderer>().material.SetTexture("_UnlitColorMap", cam.targetTexture);
-            }
-            else
-            {
-                cameraPlane.GetComponent<MeshRenderer>().material.SetTexture("_UnlitColorMap", CameraManager.EmptyTexture);
-            }
         }
     }
 }